+++ src/metrics.rs
@@ -1,30 +1,30 @@
 use anyhow::{anyhow, Result};
 use std::{
   collections::HashMap,
-  sync::{Arc, Mutex},
+  sync::{Arc, RwLock},
 };
 
 #[derive(Debug, Clone)]
 pub struct Metrics {
-  data: Arc<Mutex<HashMap<String, i64>>>,
+  data: Arc<RwLock<HashMap<String, i64>>>,
 }
 
 impl Metrics {
   pub fn new() -> Self {
     Metrics {
-      data: Arc::new(Mutex::new(HashMap::new())),
+      data: Arc::new(RwLock::new(HashMap::new())),
     }
   }
 
   pub fn inc(&self, key: impl Into<String>) -> Result<()> {
-    let mut data = self.data.lock().map_err(|e| anyhow!(e.to_string()))?;
+    let mut data = self.data.write().map_err(|e| anyhow!(e.to_string()))?;
     let counter = data.entry(key.into()).or_insert(0);
     *counter += 1;
     Ok(())
   }
 
   pub fn dec(&self, key: impl Into<String>) -> Result<()> {
-    let mut data = self.data.lock().map_err(|e| anyhow!(e.to_string()))?;
+    let mut data = self.data.write().map_err(|e| anyhow!(e.to_string()))?;
     let counter = data.entry(key.into()).or_insert(0);
     *counter -= 1;
     Ok(())
@@ -34,7 +34,7 @@ impl Metrics {
     Ok(
       self
         .data
-        .lock()
+        .read()
         .map_err(|e| anyhow!(e.to_string()))?
         .clone(),
     )